feat: add calendar to overlay
