+++ libs/shared/common/src/lib/components/datepicker/calendar.component.html
@@ -1,0 +1,59 @@
+<div class="grid gap-4 et-card bg-color-codemirror-gutters-bg">
+  <div class="flex items-center justify-between">
+    <span class="flex-grow font-bold cursor-default">{{ view() | date: 'MMMM YYYY' }}</span>
+    <div class="flex items-center gap-1">
+      <button class="w-8 et-button" (click)="changeMonth(-1)" [expensesTrackerTooltip]="previousMonthTooltip">
+        <span class="material-icons">chevron_left</span>
+      </button>
+      <button class="w-8 et-button" (click)="resetDate()" [expensesTrackerTooltip]="resetTooltip">
+        <span class="material-icons">radio_button_checked</span>
+      </button>
+      <button class="w-8 et-button" (click)="changeMonth(1)" [expensesTrackerTooltip]="nextMonthTooltip">
+        <span class="material-icons">chevron_right</span>
+      </button>
+    </div>
+  </div>
+
+  <div class="flex items-center justify-center gap-1">
+    <ng-container *ngFor="let day of daysOfWeek">
+      <button class="w-8 pointer-events-none et-button">
+        <span class="font-bold text-center">
+          {{ day }}
+        </span>
+      </button>
+    </ng-container>
+  </div>
+
+  <div class="grid gap-1 place-content-center">
+    <ng-container *ngFor="let week of calendarDays(); trackBy: weekTracker">
+      <div class="flex items-center justify-center gap-1">
+        <ng-container *ngFor="let day of week; trackBy: dayTracker">
+          <button
+            class="w-8 et-button bg-color-transparent"
+            [ngClass]="{
+              'bg-color-accent-fg text-color-fg-on-emphasis': compareWithSelectedDate(day)
+            }"
+            (click)="selectDate(day)">
+            <span
+              class="text-center"
+              [ngClass]="{
+                'text-color-fg-muted': day.getMonth() !== view().getMonth()
+              }">
+              {{ day.getDate() }}
+            </span>
+          </button>
+        </ng-container>
+      </div>
+    </ng-container>
+  </div>
+</div>
+
+<expenses-tracker-tooltip #previousMonthTooltip>
+  <span>Go to the previous month</span>
+</expenses-tracker-tooltip>
+<expenses-tracker-tooltip #resetTooltip>
+  <span>Go to today</span>
+</expenses-tracker-tooltip>
+<expenses-tracker-tooltip #nextMonthTooltip>
+  <span>Go to the next month</span>
+</expenses-tracker-tooltip>
